Update preprocessing.py
Added readManifest() method to read the file names from "tfidf_docs.txt" for use in the createNewFile() method. "tfidf_docs.txt" is the intended input to this program, it is a manifest of file names for the individual documents which need to be cleaned (test1.txt, test2.txt for testing) (doc1.txt, doc2.txt for grading). Also added a main method which uses the manifest to create the file name inputs for the createNewFile() method. Main() will run cleaning on all files in the manifest, then create a preproc_{filename}.txt for each file.